requirements.txt
Flask==2.0.2
flask-cors
pandas==1.3.5
modlamp
biopython
numpy

